+++ build.gradle
@@ -4,6 +4,8 @@ plugins {
     alias(libs.plugins.hangar)
     alias(libs.plugins.modrinth)
 
+    alias(libs.plugins.shadowjar)
+
     id 'java-library'
 }
