+++ gradle/libs.versions.toml
@@ -1,6 +1,8 @@
 [versions]
 kyori_version = "4.14.0"
 
+cluster_version = "3.1"
+
 [plugins]
 hangar = { id = "io.papermc.hangar-publish-plugin", version = "0.1.0" }
 modrinth = { id = "com.modrinth.minotaur", version = "2.8.4" }
@@ -12,6 +14,9 @@ shadowjar = { id = "com.github.johnrengelman.shadow", version = "8.1.1" }
 [libraries]
 adventure = { module = "net.kyori:adventure-api", version.ref = "kyori_version" }
 minimessage = { module = "net.kyori:adventure-text-minimessage", version.ref = "kyori_version" }
+
+cluster_paper = { module = "com.ryderbelserion.cluster.paper:cluster-paper", version.ref = "cluster_version" }
+cluster_api = { module = "com.ryderbelserion.cluster.api:cluster-api", version.ref = "cluster_version" }
 
 nbtapi = { module = "de.tr7zw:item-nbt-api", version = "2.12.0" }
