+++ paper/src/main/java/com/badbones69/crazycrates/paper/api/EventLogger.java
@@ -9,7 +9,9 @@ import org.jetbrains.annotations.NotNull;
 import us.crazycrew.crazycrates.api.enums.types.KeyType;
 import com.badbones69.crazycrates.paper.api.FileManager.Files;
 import us.crazycrew.crazycrates.paper.other.MsgUtils;
-import java.io.*;
+import java.io.BufferedWriter;
+import java.io.FileWriter;
+import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.logging.Level;
