--- paper/src/main/java/com/badbones69/crazycrates/paper/api/managers/AbstractCrate.java
@@ -1,3 +1,0 @@
-package com.badbones69.crazycrates.paper.api.managers;
-
-public abstract class AbstractCrate {}
