+++ paper/src/main/java/us/crazycrew/crazycrates/paper/listeners/crates/QuadCrateListener.java
@@ -9,8 +9,8 @@ import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.player.PlayerTeleportEvent;
 import org.bukkit.inventory.meta.ItemMeta;
 import us.crazycrew.crazycrates.paper.CrazyCrates;
-import com.badbones69.crazycrates.paper.api.managers.QuadCrateManager;
-import com.badbones69.crazycrates.paper.api.managers.quadcrates.SessionManager;
+import us.crazycrew.crazycrates.paper.managers.crates.other.quadcrates.QuadCrateManager;
+import us.crazycrew.crazycrates.paper.managers.crates.other.quadcrates.SessionManager;
 import com.badbones69.crazycrates.paper.api.objects.Crate;
 import com.badbones69.crazycrates.paper.api.objects.ItemBuilder;
 import com.badbones69.crazycrates.paper.api.objects.Prize;
