+++ paper/src/main/java/us/crazycrew/crazycrates/paper/managers/crates/other/AbstractCrateManager.java
@@ -1,0 +1,3 @@
+package us.crazycrew.crazycrates.paper.managers.crates.other;
+
+public abstract class AbstractCrateManager {}
