+++ paper/src/main/java/us/crazycrew/crazycrates/paper/managers/crates/other/quadcrates/SessionManager.java
@@ -1,6 +1,5 @@
-package com.badbones69.crazycrates.paper.api.managers.quadcrates;
+package us.crazycrew.crazycrates.paper.managers.crates.other.quadcrates;
 
-import com.badbones69.crazycrates.paper.api.managers.QuadCrateManager;
 import org.bukkit.entity.Player;
 
 public class SessionManager {
