+++ paper/src/main/java/us/crazycrew/crazycrates/paper/managers/crates/types/CosmicCrate.java
@@ -1,6 +1,6 @@
 package us.crazycrew.crazycrates.paper.managers.crates.types;
 
-import com.badbones69.crazycrates.paper.api.managers.CosmicCrateManager;
+import us.crazycrew.crazycrates.paper.managers.crates.other.CosmicCrateManager;
 import com.badbones69.crazycrates.paper.api.objects.Crate;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
